Fix missing import

+++ src/pyread_swift/read_header.py
@@ -70,7 +70,7 @@ def read_header(params):
     if "NumPart_Total_HighWord" in header.keys():
         for i in range(len(header["NumPart_Total"])):
             header["NumPart_Total"][i] = header["NumPart_Total"][i] + (
-                header["NumPart_Total_HighWord"][i] * 2**32
+                header["NumPart_Total_HighWord"][i].astype(np.uint64) * 2**32
             )
             header["NumPart_Total_HighWord"][i] = 0
